database-service/app/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
alembic==1.13.1
PyYAML==6.0.1 